requirements.txt
torch>=2.0.0
torchvision>=0.15.0
ultralytics>=8.0.0
segmentation-models-pytorch>=0.3.3
timm>=0.9.0

# Geospatial Libraries
rasterio>=1.3.0
gdal>=3.5.0
geopandas>=0.13.0
shapely>=2.0.0
pyproj>=3.5.0
fiona>=1.9.0

# Computer Vision
opencv-python>=4.7.0
albumentations>=1.3.0
scikit-image>=0.20.0

# Data Science
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.2.0
scipy>=1.10.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.14.0
folium>=0.14.0
bokeh>=3.1.0

# Jupyter and Development
jupyter>=1.0.0
notebook>=6.5.0
ipywidgets>=8.0.0

# Utils and Performance
tqdm>=4.65.0
pillow>=9.5.0
imageio>=2.28.0
joblib>=1.2.0
psutil>=5.9.0

# Documentation
sphinx>=6.0.0
sphinx-rtd-theme>=1.2.0

# Configuration
PyYAML>=6.0

# Testing
pytest>=7.3.0
pytest-cov>=4.0.0
